% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eds221-day5-eod-function.R
\name{baking}
\alias{baking}
\title{What I Baked Today}
\usage{
baking(total_baked, baked_good)
}
\arguments{
\item{total_baked}{Total amount of baked items produced}

\item{baked_good}{The item baked}
}
\description{
What I Baked Today
}
\examples{
baking(15, "chocolate chip cookies")
}
