% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eds221-day5-eod-function.R
\name{lactose_intolerant}
\alias{lactose_intolerant}
\title{A Measure of Lactose Intolerance}
\usage{
lactose_intolerant(dairy_eaten, minutes)
}
\arguments{
\item{dairy_eaten}{The type of dairy consumed}

\item{minutes}{The amount of time passed (in minutes) since dairy was consumed}
}
\description{
A Measure of Lactose Intolerance
}
\examples{
lactose_intolerant("brie", 30)
}
