% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pussinboots.R
\name{puss_in_boots}
\alias{puss_in_boots}
\title{A function to report Puss in Boot's meals}
\usage{
puss_in_boots(total_eaten, fish_type, location)
}
\arguments{
\item{total_eaten}{total number of fish Puss in Boots has eaten}

\item{fish_type}{what kind of fish species Puss in Boots is eating}

\item{location}{where Puss in Boots is eating his fish}
}
\description{
A function to report Puss in Boot's meals
}
\examples{
puss_in_boots(30, "salmon", "long beach")
}
